facts: consolidate the pipeline into one driver
derive_profile.py (61 lines, one consumer), make_records.py (records
stage), and golden_snapshots.py (freshness) merge into gen_facts.py as
subcommands (--records, --snapshot-update/--snapshot-check). The module
count drops 12 -> 7; what remains separate has a hard boundary: the three
probes execute inside different framework containers, sitecustomize's
filename IS the injection mechanism, gen_dummy_models is a rarely-run
builder, build_images.sh is shell. Behavior-preserving: --records
regenerates all 335 records byte-identical (modulo the new ep/dp/a2a
runtime fields), snapshots re-render surface-identical, coverage passes.
